+++ admin/class-sessions-admin.php
@@ -296,6 +296,7 @@ class Sessions_Admin {
 				Option::network_set( 'use_cdn', array_key_exists( 'pose_plugin_options_usecdn', $_POST ) ? (bool) filter_input( INPUT_POST, 'pose_plugin_options_usecdn' ) : false );
 				Option::network_set( 'display_nag', array_key_exists( 'pose_plugin_options_nag', $_POST ) ? (bool) filter_input( INPUT_POST, 'pose_plugin_options_nag' ) : false );
 				Option::network_set( 'analytics', array_key_exists( 'pose_plugin_features_analytics', $_POST ) ? (bool) filter_input( INPUT_POST, 'pose_plugin_features_analytics' ) : false );
+				Option::network_set( 'forceip', array_key_exists( 'pose_plugin_features_forceip', $_POST ) ? (bool) filter_input( INPUT_POST, 'pose_plugin_features_forceip' ) : false );
 				Option::network_set( 'history', array_key_exists( 'pose_plugin_features_history', $_POST ) ? (string) filter_input( INPUT_POST, 'pose_plugin_features_history', FILTER_SANITIZE_NUMBER_INT ) : Option::network_get( 'history' ) );
 				Option::network_set( 'rolemode', array_key_exists( 'pose_plugin_features_rolemode', $_POST ) ? (string) filter_input( INPUT_POST, 'pose_plugin_features_rolemode', FILTER_SANITIZE_NUMBER_INT ) : Option::network_get( 'rolemode' ) );
 				$message = esc_html__( 'Plugin settings have been saved.', 'sessions' );
@@ -475,6 +476,22 @@ class Sessions_Admin {
 			]
 		);
 		register_setting( 'pose_plugin_features_section', 'pose_plugin_features_rolemode' );
+		add_settings_field(
+			'pose_plugin_features_forceip',
+			esc_html__( 'IP detection', 'sessions' ),
+			[ $form, 'echo_field_checkbox' ],
+			'pose_plugin_features_section',
+			'pose_plugin_features_section',
+			[
+				'text'        => esc_html__( 'Override WordPress', 'sessions' ),
+				'id'          => 'pose_plugin_features_forceip',
+				'checked'     => Option::network_get( 'forceip' ),
+				'description' => esc_html__( 'If checked, Sessions will improve IP detection used for tokens (recommended).', 'sessions' ),
+				'full_width'  => false,
+				'enabled'     => true,
+			]
+		);
+		register_setting( 'pose_plugin_features_section', 'pose_plugin_features_forceip' );
 		add_settings_field(
 			'pose_plugin_features_analytics',
 			esc_html__( 'Analytics', 'sessions' ),