+++ includes/system/class-session.php
@@ -493,6 +493,55 @@ class Session {
 		if ( -1 === (int) Option::network_get( 'rolemode' ) ) {
 			return $user;
 		}
+		if ( $user instanceof \WP_Error ) {
+			if ( empty( $username ) && empty( $password ) ) {
+				//$a = wp_validate_auth_cookie( $_COOKIE[ LOGGED_IN_COOKIE ], 'logged_in' );
+				//Logger::critical($a?$a:'false');
+
+
+
+				/*$u = wp_authenticate_cookie( $user, $username, $password );
+				if ( $u instanceof \WP_Error ) {
+					Logger::critical('User id error');
+					Logger::critical($u->get_error_message(), $u->get_error_code());
+					return $user;
+				}
+				$user = $u;*/
+
+
+				/*
+
+				$user_id = wp_validate_auth_cookie();
+				if ( $user_id ) {
+					$u = new \WP_User( $user_id );
+					if ( $u instanceof \WP_Error ) {
+						Logger::critical('User id error');
+						Logger::critical($u->get_error_message(), $u->get_error_code());
+						return $user;
+					}
+					$user = $u;
+				} else {
+					Logger::critical('No user id');
+					return $user;
+				}*/
+/*
+				global $auth_secure_cookie;
+
+				if ( $auth_secure_cookie ) {
+					$auth_cookie = SECURE_AUTH_COOKIE;
+				} else {
+					$auth_cookie = AUTH_COOKIE;
+				}
+
+				if ( ! empty( $_COOKIE[ $auth_cookie ] ) ) {
+					return new WP_Error( 'expired_session', __( 'Please log in again.' ) );
+				}
+*/
+				// If the cookie is not set, be silent.
+			}
+		}
+
+
 		if ( $user instanceof \WP_User ) {
 			$this->user_id  = $user->ID;
 			$this->user     = $user;
@@ -724,6 +773,9 @@ class Session {
 	 * @since    1.0.0
 	 */
 	public static function init() {
+		if ( Option::network_get( 'forceip' ) ) {
+			$_SERVER['REMOTE_ADDR'] = IP::get_current();
+		}
 		if ( ! isset( self::$instance ) ) {
 			self::$instance = new static();
 		}